requirements.txt
# JR Manufacturing Smart Dashboard - Requirements
# Modular predictive maintenance system

# Core framework
streamlit>=1.28.0

# Data processing
pandas>=1.5.0
numpy>=1.24.0

# Visualization
plotly>=5.15.0

# Machine Learning
scikit-learn>=1.3.0
joblib>=1.3.0

# Configuration
PyYAML>=6.0

# Database
sqlite3  # Built-in with Python

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3

# Optional: Advanced ML (uncomment if needed)
# torch>=2.0.0
# tensorflow>=2.13.0
# xgboost>=1.7.0